Move static methods in translator to utils
